Add files via upload
Bus Pass Management Project in PHP is a web-based technology that will manage the records of the pass which is issued by the administrative and also help to provide online bus pass to people who need to travel daily. The Bus Pass Management System project is helpful to bus administration by reducing the paperwork, and time consumption and making the process of getting a bus pass simple and fast.